moved StepRepository to Database folder
