+++ app/src/main/java/com/github/k409/fitflow/ui/step_counter/StepCounterWorker.kt
@@ -7,6 +7,7 @@ import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.github.k409.fitflow.DataModels.Step
+import com.github.k409.fitflow.Database.StepRepository
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import java.time.LocalDate
